refactor: update BusinessException import paths and enhance Company entity validation
